Split non-UI code into lib crate
Move existing code into `crates/app` and create a secondary
`crates/lib` to hold all non-UI logic.
