+++ crates/app/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "egui-trace"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lib = { path = "../lib" }
+eframe = { version = "0.21.0", features = ["dark-light"] }
+egui_dock = "0.4.2"
+egui_extras = "0.21.0"
+rfd = "0.11.3"
+tokio = { version = "1.27.0", features = ["full"] }
+tracing = "0.1.37"
+tracing-subscriber = "0.3.16"
