+++ crates/lib/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "lib"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+axum = "0.6.16"
+axum-extra = { version = "0.7.4", features = ["protobuf"] }
+chrono = { version = "0.4", features = ["serde"] }
+prost = "0.11.9"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.27.0", features = ["full"] }
+tracing = "0.1.37"
+
+[build-dependencies]
+prost-build = "0.11.9"
