+++ crates/lib/build.rs
