+++ crates/lib/opentelemetry-proto/.github/CODEOWNERS
@@ -1,0 +1,25 @@
+#####################################################
+#
+# List of approvers for OpenTelemetry Proto repository
+#
+#####################################################
+#
+# Learn about membership in OpenTelemetry community:
+#  https://github.com/open-telemetry/community/blob/main/community-membership.md
+# 
+#
+# Learn about CODEOWNERS file format: 
+#  https://help.github.com/en/articles/about-code-owners
+#
+
+# Global owners, will be the owners for everything in the repo. Membership is tracked via https://github.com/open-telemetry/community/blob/main/community-members.md 
+*   @open-telemetry/specs-approvers
+
+# Trace owners (global + trace)
+opentelemetry/proto/trace/    @open-telemetry/specs-approvers @open-telemetry/specs-trace-approvers
+
+# Metrics owners (global + metrics)
+opentelemetry/proto/metrics/  @open-telemetry/specs-approvers @open-telemetry/specs-metrics-approvers
+
+# Logs owners (global + logs)
+opentelemetry/proto/logs/     @open-telemetry/specs-approvers @open-telemetry/specs-logs-approvers
