+++ crates/lib/opentelemetry-proto/.github/workflows/auto-assign-owners.yml
@@ -1,0 +1,14 @@
+name: 'Auto Assign'
+on:
+  pull_request_target:
+    types: [opened, ready_for_review]
+
+jobs:
+  add-owner:
+    runs-on: ubuntu-latest
+    steps:
+      - name: run
+        uses: kentaro-m/auto-assign-action@v1.1.2
+        with:
+          configuration-path: ".github/auto_assign.yml"
+          repo-token: '${{ secrets.GITHUB_TOKEN }}'
