+++ crates/lib/opentelemetry-proto/.gitignore
@@ -1,0 +1,22 @@
+# IntelliJ IDEA
+.idea
+*.iml
+
+# VS Code
+.vscode
+.classpath
+.project
+.settings/
+
+# OS X
+.DS_Store
+
+# Emacs
+*~
+\#*\#
+
+# Vim
+.swp
+
+# Generated code
+/gen/
