+++ crates/lib/opentelemetry-proto/buf.yaml
@@ -1,0 +1,6 @@
+version: v1
+
+# See https://docs.buf.build/breaking/configuration
+breaking:
+  use:
+    - WIRE
