+++ crates/lib/opentelemetry-proto/opentelemetry/proto/collector/logs/v1/logs_service_http.yaml
@@ -1,0 +1,9 @@
+# This is an API configuration to generate an HTTP/JSON -> gRPC gateway for the
+# OpenTelemetry service using github.com/grpc-ecosystem/grpc-gateway.
+type: google.api.Service
+config_version: 3
+http:
+ rules:
+ - selector: opentelemetry.proto.collector.logs.v1.LogsService.Export
+   post: /v1/logs
+   body: "*"
