--- src/collector/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod panel;
-pub(crate) mod server;
